requirements.txt
beautifulsoup4~=4.11.1
requests~=2.28.1
pandas~=1.5.3
SQLAlchemy~=2.0.0
python-dotenv~=0.21.1
python-dateutil~=2.8.2